Nathan: debug graph error, training seems to be working

+++ lib_dtypes.py
@@ -2,6 +2,6 @@
 import torch
 
 #
-DTYPE_FLOAT: torch.dtype = torch.float16
+DTYPE_FLOAT: torch.dtype = torch.float32
 DTYPE_INT: torch.dtype = torch.int64
